requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 缓存
redis==5.0.1

# 配置管理
python-dotenv==1.0.0
pydantic-settings==2.0.3

# 安全认证
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AI - Langchain 体系
langchain==1.0.4
langgraph==1.0.2
langchain_community==0.4.1
langchain-openai>=0.1.0

# AI - Claude SDK 体系
claude-agent-sdk>=0.1.0  # Claude SDK Agent 支持